# openAPI.yaml
...
openapi: 3.0.0
info:
  title: OpenCart Monolithic Backend API
  version: 1.0.0
  description: API specification for the OpenCart monolithic backend, handling products, users, and orders/transactions.
servers:
  - url: http://localhost:5000/api
    description: Local Development Server (via kubectl port-forward)
  # - url: http://opencart.local/api # Alternative local access via Ingress
  # - url: https://your-cloud-domain.com/api # Placeholder for Cloud deployment
...
paths:
...
  /user/register:
    post:
      tags:
        - Users
      summary: Register a new user
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserInput'
      responses:
        '201':
          description: User registered successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/User'
        '400':
          description: Invalid input or email already exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
                example:
                  success: false
                  message: Email is already linked to an account.
        '500':
          $ref: '#/components/responses/ServerError'
  /user/login:
    post:
      tags:
        - Users
      summary: User login
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  format: password
                  example: mySecurePassword123!
      responses:
        '200':
          description: User logged in successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Invalid credentials.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
                example:
                  success: false
                  message: Invalid email or password.
        '500':
          $ref: '#/components/responses/ServerError'
...
  /orders/createOrder:
    post:
      tags:
        - Orders & Transactions
      summary: Create a new order (Transaction)
      operationId: createOrder
      # security:
      #   - bearerAuth: [] # Requires authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionInput'
      responses:
        '201':
          description: Order/Transaction created successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Transaction'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          description: User or Product not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
                example:
                  success: false
                  message: User not found.
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  schemas:
...
    User:
      type: object
      properties:
        _id:
          type: string
          example: 60c72b2f9b1e8e001c8e0002
        firstName:
          type: string
          example: Demo
        lastName:
          type: string
          example: User
        email:
          type: string
          format: email
          example: demo.user@example.com
        phone:
          type: string
          example: "1234567890"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - firstName
        - lastName
        - email
        - password # Note: Password is usually not returned in GET responses for security
    UserInput:
      type: object
      properties:
        firstName:
          type: string
          example: Demo
        lastName:
          type: string
          example: User
        email:
          type: string
          format: email
          example: demo.user@example.com
        password:
          type: string
          format: password
          example: DemoPassword123!
        phone:
          type: string
          example: "1234567890"
      required:
        - firstName
        - lastName
        - email
        - password
    Transaction:
      type: object
      properties:
        _id:
          type: string
          example: 60c72b2f9b1e8e001c8e0003
        user:
          type: string
          example: 60c72b2f9b1e8e001c8e0002 # User ID
        productId:
          type: string
          example: 60c72b2f9b1e8e001c8e0001 # Product ID
        quantity:
          type: integer
          example: 1
        totalPrice:
          type: number
          format: float
          example: 199.99
        currency:
          type: string
          enum: ['USD', 'EUR', 'GBP', 'JPY', 'AUD', 'CAD', 'CHF', 'CNY', 'SEK', 'NZD','KES', 'NGN', 'GHS', 'ZAR', 'INR', 'PKR', 'BDT', 'LKR', 'MUR', 'TND', 'MAD', 'EGP']
          example: USD
        transactionDate:
          type: string
          format: date-time
        status:
          type: string
          enum: ['pending', 'completed', 'failed', 'refunded', 'cancelled', 'processing', 'on_hold']
          example: processing
        paymentMethod:
          type: string
          enum: ['credit_card', 'paypal', 'bank_transfer', 'crypto']
          example: credit_card
        shippingAddress:
          type: string
          example: 123 Main Street, Anytown, 12345, USA
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - user
        - productId
        - quantity
        - totalPrice
        - currency
        - paymentMethod
        - shippingAddress
    TransactionInput:
      type: object
      properties:
        user:
          type: string
          description: ID of the user placing the order.
          example: 60c72b2f9b1e8e001c8e0002
        productId:
          type: string
          description: ID of the product being ordered.
          example: 60c72b2f9b1e8e001c8e0001
        quantity:
          type: integer
          example: 1
        totalPrice:
          type: number
          format: float
          example: 199.99
        currency:
          type: string
          enum: ['USD', 'EUR', 'GBP', 'JPY', 'AUD', 'CAD', 'CHF', 'CNY', 'SEK', 'NZD','KES', 'NGN', 'GHS', 'ZAR', 'INR', 'PKR', 'BDT', 'LKR', 'MUR', 'TND', 'MAD', 'EGP']
          example: USD
        paymentMethod:
          type: string
          enum: ['credit_card', 'paypal', 'bank_transfer', 'crypto']
          example: credit_card
        shippingAddress:
          type: string
          example: 123 Main Street, Anytown, 12345, USA
      required:
        - user
        - productId
        - quantity
        - totalPrice
        - currency
        - paymentMethod
        - shippingAddress
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Something went wrong.
  responses:
    BadRequest:
      description: Invalid input provided.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
            example:
              success: false
              message: Invalid input.
...
    ServerError:
      description: Internal server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
            example:
              success: false
              message: Internal server error.
  securitySchemes:
    bearerAuth: # Can be used for JWT authentication
      type: http
      scheme: bearer
      bearerFormat: JWT